gh-1344 Fixed all tests and other build schemes
