--- Multisig/Cross-layer/Configuration/Staging.Mainnet.xcconfig
@@ -1,15 +1,0 @@
-//
-//  Staging.xcconfig
-//  Multisig
-//
-//  Created by Dmitry Bespalov on 27.05.20.
-//  Copyright © 2020 Gnosis Ltd. All rights reserved.
-//
-
-// Configuration settings file format documentation can be found at:
-// https://help.apple.com/xcode/#/dev745c5c974
-
-#include "Config.xcconfig"
-
-SERVICE_ENV = STAGING
-NETWORK = MAINNET
