+++ NotificationServiceExtension/NotificationService.Config.xcconfig
@@ -12,16 +12,10 @@
 // DEV | STAGING | PROD
 SERVICE_ENV = STAGING
 
-// RINKEBY | MAINNET
-NETWORK = RINKEBY
-
 // Bundle Identifier
-_APP_BUNDLE_ID_PROD_MAINNET    = io.gnosis.multisig.prod.mainnet.NotificationServiceExtension
-_APP_BUNDLE_ID_PROD_RINKEBY    = io.gnosis.multisig.prerelease.NotificationServiceExtension
-_APP_BUNDLE_ID_STAGING_MAINNET = io.gnosis.multisig.staging.mainnet.NotificationServiceExtension
-_APP_BUNDLE_ID_STAGING_RINKEBY = io.gnosis.multisig.staging.rinkeby.NotificationServiceExtension
-_APP_BUNDLE_ID_DEV_MAINNET     = io.gnosis.multisig.dev.mainnet.NotificationServiceExtension
-_APP_BUNDLE_ID_DEV_RINKEBY     = io.gnosis.multisig.dev.rinkeby.NotificationServiceExtension
+_APP_BUNDLE_ID_PROD     = io.gnosis.multisig.prod.mainnet.NotificationServiceExtension
+_APP_BUNDLE_ID_STAGING  = io.gnosis.multisig.staging.mainnet.NotificationServiceExtension
+_APP_BUNDLE_ID_DEV      = io.gnosis.multisig.dev.mainnet.NotificationServiceExtension
 
 // Xcode Build Settings
-PRODUCT_BUNDLE_IDENTIFIER           = $(_APP_BUNDLE_ID_$(SERVICE_ENV)_$(NETWORK):default=$(_APP_BUNDLE_ID_PROD_MAINNET))
+PRODUCT_BUNDLE_IDENTIFIER           = $(_APP_BUNDLE_ID_$(SERVICE_ENV):default=$(_APP_BUNDLE_ID_PROD))
