+++ NotificationServiceExtension/NotificationService.Production.xcconfig
@@ -1,8 +1,8 @@
 //
-//  Development.Rinkeby.xcconfig
+//  ProductionMainnet.xcconfig
 //  Multisig
 //
-//  Created by Dmitry Bespalov on 28.05.20.
+//  Created by Dmitry Bespalov on 27.05.20.
 //  Copyright © 2020 Gnosis Ltd. All rights reserved.
 //
 
@@ -11,5 +11,4 @@
 
 #include "NotificationService.Config.xcconfig"
 
-SERVICE_ENV = DEV
-NETWORK = RINKEBY
+SERVICE_ENV = PROD
