+++ Shared/CoreData/CoreDataProtocol.swift
