+++ Shared/CoreData/Safe+CoreDataClass.swift
