+++ Shared/CoreData/Safe+CoreDataProperties.swift
