+++ Shared/Utils/CharacterSet+Hex.swift
